Changes for CW Pro5.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@2354 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csterr/ddghemap.cpp
@@ -318,7 +318,7 @@ void ddgHeightMap::canyonize(float f )
 		for (unsigned c = 0; c < _cols; c++ )
 		{
 			unsigned int d = get(r,c);
-            unsigned int d2 = (unsigned int)ddgUtil::clamp(pow(d,f),0,0xFFFF);
+            unsigned int d2 = (unsigned int)ddgUtil::clamp(pow(d,(unsigned int)f),0,0xFFFF);
 			set(r,c,d2);
 		}
 }
